requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
xgboost
catboost
dill
azure.ai.ml
mlflow
psutil
azureml-mlflow
tqdm
fsspec
#-e .